[FIX] stock_release_channel: A released picking is a picking with a last_release_date
Before this change a released picking was a picking that do no more need to be released. With the management of backorders introduced in https://github.com/OCA/wms/pull/478, some released pickings could still need to be released since qties were not available for some product at release time. Fortunately, at the same time as this change in the management of the release process, a new field has been added to log the timestamp of the last release done on a picking. Therefore, to kown if picking is released we can replace the search on 'need_release' by a search on 'last_release_date'. A released picking is a picking with a last_release_date filled.